requirements.txt
streamlit
pandas
plotly
scikit-learn
reportlab
python-dotenv
